quiz back end in progress

+++ question-service/src/main/java/com/redhun/question_service/dto/QuestionRequest.java
@@ -46,7 +46,7 @@ public class QuestionRequest {
         this.difficultyId = difficultyId;
     }
 
-    public int getQuestionType() {
+    public long getQuestionType() {
         return questionType;
     }
